// src/utils/enums.ts
export enum KeyboardLayoutNames {
  MACOS_US_QWERTY = "MACOS_US_QWERTY",
  MACOS_US_DVORAK = "MACOS_US_DVORAK",
  MACOS_US_COLEMAK = "MACOS_US_COLEMAK",
  MACOS_FR_AZERTY = "MACOS_FR_AZERTY",
  MACOS_DE_QWERTZ = "MACOS_DE_QWERTZ",
  MACOS_ES_QWERTY = "MACOS_ES_QWERTY",
}

export enum Levels {
  LEVEL_1 = "1",
  LEVEL_2 = "2",
  LEVEL_3 = "3",
}

export enum Languages {
  ENGLISH = "en",
  FRENCH = "fr",
  GERMAN = "de",
  SPANISH = "es",
}